wip: accumulateOrderFills
